+++ projects/cas3-core/Cargo.toml
@@ -6,7 +6,7 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-cairo-rs = { version = "0.18.2", features = ["svg"] }
+#cairo-rs = { version = "0.18.2", features = ["svg"] }
 num-bigint = "0.4.4"
 num-traits = "0.2.17"
 ordered-float = "4.1.0"
